@@ -1,20 +1,21 @@
 import React from 'react'
-import ReactDOM from 'react-dom'
-import reportWebVitals from './reportWebVitals'
+import ReactDOM from 'react-dom/client';
+import reportWebVitals from './reportWebVitals';
 
-import 'antd/dist/antd.css'
+import 'antd/dist/reset.css';
+
 
 import App from './App'
 import './Style/Main.scss'
 
-ReactDOM.render(
+const root = ReactDOM.createRoot(document.getElementById('root'));
+root.render(
     <React.StrictMode>
         <App />
-    </React.StrictMode>,
-    document.getElementById('root')
-)
+    </React.StrictMode>
+);
 
 // If you want to start measuring performance in your app, pass a function
 // to log results (for example: reportWebVitals(console.log))
 // or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-reportWebVitals()
+reportWebVitals();
